Move scripts from processing to datasets folder
